feat: include /scope export
