+++ package.json
@@ -24,6 +24,12 @@
       "import": "./ssr.mjs",
       "require": "./ssr.js",
       "default": "./ssr.mjs"
+    },
+    "./scope.mjs": "./scope.mjs",
+    "./scope": {
+      "import": "./scope.mjs",
+      "require": "./scope.js",
+      "default": "./scope.mjs"
     }
   },
   "main": "reflect.cjs.js",
@@ -34,6 +40,7 @@
     "core",
     "index.d.ts",
     "ssr.d.ts",
+    "scope.d.ts",
     "reflect.cjs.js",
     "reflect.cjs.js.map",
     "reflect.mjs",
@@ -41,7 +48,11 @@
     "ssr.js",
     "ssr.js.map",
     "ssr.mjs",
-    "ssr.mjs.map"
+    "ssr.mjs.map",
+    "scope.js",
+    "scope.js.map",
+    "scope.mjs",
+    "scope.mjs.map"
   ],
   "scripts": {
     "test:code": "jest ./src",
@@ -50,7 +61,7 @@
     "test:code:build": "jest ./dist-test",
     "test-build": "yarn test:code:build",
     "build": "yarn clear-build && yarn rollup --config ./rollup.config.js && cp index.d.ts reflect.d.ts",
-    "clear-build": "rm -f index.* && rm -f reflect.* && rm -f ssr.* && rm -rf dist && rm -rf core",
+    "clear-build": "rm -f index.* && rm -f reflect.* && rm -f ssr.* && rm -f scope.* && rm -rf dist && rm -rf core",
     "prepublishOnly": "yarn build"
   },
   "devDependencies": {
