+++ rollup.config.js
@@ -46,6 +46,7 @@ const plugins = (isEsm) => [
 
 const noSsr = './src/index.ts';
 const ssr = './src/ssr.ts';
+const scope = './src/scope.ts';
 const external = [
   'effector',
   'effector/effector.mjs',
@@ -79,6 +80,17 @@ export default [
     },
   },
   {
+    input: scope,
+    external,
+    plugins: plugins(true),
+    output: {
+      file: './scope.mjs',
+      format: 'es',
+      sourcemap: true,
+      externalLiveBindings: false,
+    },
+  },
+  {
     input: noSsr,
     external,
     plugins: plugins(),
@@ -97,6 +109,19 @@ export default [
     plugins: plugins(),
     output: {
       file: './ssr.js',
+      format: 'cjs',
+      freeze: false,
+      exports: 'named',
+      sourcemap: true,
+      externalLiveBindings: false,
+    },
+  },
+  {
+    input: scope,
+    external,
+    plugins: plugins(),
+    output: {
+      file: './scope.js',
       format: 'cjs',
       freeze: false,
       exports: 'named',
