+++ src/ssr.ts
@@ -1,14 +1,4 @@
-import * as effectorReactSSR from 'effector-react/ssr';
-import {
-  variantFactory,
-  reflectCreateFactory,
-  reflectFactory,
-  listFactory,
-} from './core';
-
-export const reflect = reflectFactory(effectorReactSSR);
-export const createReflect = reflectCreateFactory(effectorReactSSR);
-
-export const variant = variantFactory(effectorReactSSR);
-
-export const list = listFactory(effectorReactSSR);
+console.error(
+  '`@effector/reflect/ssr` is deprecated, import from `@effector/reflect/scope` instead',
+);
+export * from './scope';
